refactor: move code with private env to different file
